refactor(continuity): move is_substantive and resolvePredecessor into _lib
Preparation for step 1.2 of road-to-continuity-writer-activation, split
out because it changes no behaviour and every existing fixture proves it.

The Stop-path hook needs two predicates that lived in CLI modules:

  - is_substantive + SUBSTANTIVE_TOKEN_FLOOR were in
    _cli/handoff_sessions.ts, whose module graph a Stop hook has no
    business loading. They move to _lib/session_eol.ts, which already
    owns StoredEolCounters -- the type the predicate reads.
  - resolvePredecessor was in _cli/cmd_session_recycle.ts. It moves to
    _lib/recycle_envelope_paths.ts, which already owns
    recycle_consumed_rel and predecessorTracePresent, so the whole
    lineage question now lives in one file.

Both are re-exported from their former homes, so no importer changes and
the resume picker keeps naming the surface its readers know.

This is the shared-path repair rather than a second implementation: a
copy of is_substantive in the hook would be a second substantive floor
to keep in step with the one the resume picker uses, and the two
disagreeing is a session that the picker lists and the writer skips.

Verification: tsc --noEmit clean; 119 fixtures across
handoff_sessions, _lib_session_eol, session_recycle, recycle_roundtrip,
continuity_record_session_key and recycle_envelope_consumer all green
unchanged. No test was added here on purpose -- a relocation whose
existing fixtures pass through the new path IS the evidence, and a new
test asserting the move would only restate the import.

One line carries a code-comment-allow escape: the relocated
SUBSTANTIVE_TOKEN_FLOOR docblock cites its derivation evidence, and
lint_code_comments is diff-scoped so an unchanged docblock reds only
because the file it sits in is new to the diff. The escape names that
reason rather than deleting a citation whose whole point is that the
threshold may only move on cited evidence.
